+++ packages/experimental-identity-and-auth/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 0.7.16
+
+### Patch Changes
+
+- Updated dependencies [fcf1366]
+  - @smithy/core@3.31.1
+  - @smithy/signature-v4@5.6.12
+
 ## 0.7.15
 
 ### Patch Changes