+++ packages/smithy-client/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @smithy/smithy-client
 
+## 4.14.16
+
+### Patch Changes
+
+- Updated dependencies [fcf1366]
+  - @smithy/core@3.31.1
+
 ## 4.14.15
 
 ### Patch Changes