+++ packages/undici-http-handler/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 3.0.10
+
+### Patch Changes
+
+- Updated dependencies [fcf1366]
+  - @smithy/core@3.31.1
+
 ## 3.0.9
 
 ### Patch Changes